+++ attrib/interactive.c
@@ -103,7 +103,7 @@ static void events_handler(const uint8_t *pdu, uint16_t len, gpointer user_data)
 {
 	uint8_t *opdu;
 	uint16_t handle, i, olen;
-	int plen;
+	size_t plen;
 
 	handle = att_get_u16(&pdu[1]);
