+++ src/attrib-server.h
@@ -25,14 +25,15 @@
 uint16_t attrib_db_find_avail(struct btd_adapter *adapter, bt_uuid_t *svc_uuid,
 							uint16_t nitems);
 struct attribute *attrib_db_add(struct btd_adapter *adapter, uint16_t handle,
-				bt_uuid_t *uuid, int read_reqs, int write_reqs,
-				const uint8_t *value, int len);
+				bt_uuid_t *uuid, size_t read_reqs,
+				size_t write_reqs, const uint8_t *value,
+				size_t len);
 int attrib_db_update(struct btd_adapter *adapter, uint16_t handle,
 					bt_uuid_t *uuid, const uint8_t *value,
-					int len, struct attribute **attr);
+					size_t len, struct attribute **attr);
 int attrib_db_del(struct btd_adapter *adapter, uint16_t handle);
 int attrib_gap_set(struct btd_adapter *adapter, uint16_t uuid,
-						const uint8_t *value, int len);
+					const uint8_t *value, size_t len);
 uint32_t attrib_create_sdp(struct btd_adapter *adapter, uint16_t handle,
 							const char *name);
 void attrib_free_sdp(uint32_t sdp_handle);
